+++ Tests/Fixtures/Renderer/Atom/FullFeed.php
@@ -12,10 +12,12 @@ declare(strict_types=1);
 namespace Brotkrueml\FeedGenerator\Tests\Fixtures\Renderer\Atom;
 
 use Brotkrueml\FeedGenerator\Contract\AuthorInterface;
+use Brotkrueml\FeedGenerator\Contract\CategoryInterface;
 use Brotkrueml\FeedGenerator\Contract\FeedInterface;
 use Brotkrueml\FeedGenerator\Contract\ImageInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
 use Brotkrueml\FeedGenerator\Entity\Author;
+use Brotkrueml\FeedGenerator\Entity\Category;
 use Brotkrueml\FeedGenerator\Entity\Image;
 
 final class FullFeed implements FeedInterface
@@ -79,6 +81,25 @@ final class FullFeed implements FeedInterface
     public function getImage(): ?ImageInterface
     {
         return new Image('https://example.org/some-image');
+    }
+
+    /**
+     * @return CategoryInterface[]
+     */
+    public function getCategories(): array
+    {
+        return [
+            new Category(
+                'some term',
+                'https://example.org/some-term',
+                'some label',
+            ),
+            new Category(
+                'another term',
+                'https://example.org/another-term',
+                'another label',
+            ),
+        ];
     }
 
     /**