+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Substitute laminas/laminas-feed and jdecool/jsonfeed with custom implementation
 - Rename Enclosure class to Attachment
 - Allow multiple attachments for JSON feeds
+- Integrate FeedCategoryInterface into FeedInterface
 
 ## [0.4.0] - 2022-10-31